aide-backend/requirements.txt
# FastAPI Backend Requirements

# Core FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Async file operations
aiofiles==23.2.1

# HTTP client for API validation
httpx==0.25.2

# CORS middleware
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# AIDE ML dependencies (from main requirements.txt)
black==24.3.0
funcy==2.0
humanize==4.8.0
jsonschema==4.19.2
numpy==1.26.2
openai>=1.69.0
anthropic>=0.20.0
pandas==2.1.4
requests==2.32.2
scikit-learn==1.5.0
rich==13.7.0
scipy==1.11.4
dataclasses_json>=0.6.4
omegaconf>=2.3.0
loguru==0.7.2
shutup>=0.2.0
tqdm==4.66.3
coolname>=2.2.0
igraph>=0.11.3
genson>=1.2.0

# Database dependencies
sqlalchemy==2.0.23
# sqlite3 is built-in with Python

# ML libraries that AIDE might need (minimal for local dev)
lightgbm
matplotlib
seaborn
plotly
Pillow
PyYAML
h5py
